+++ LabelStoreMax/lib/app/controllers/controller.dart
@@ -8,10 +8,16 @@
 //  distributed under the License is distributed on an "AS IS" BASIS,
 //  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 
+import 'package:flutter/material.dart';
 import 'package:nylo_support/controllers/controller.dart';
 
 /// Base Controller for the Nylo
-/// See more on controllers here - https://nylo.dev/docs/1.x/controllers
+/// See more on controllers here - https://nylo.dev/docs/2.x/controllers
 class Controller extends BaseController {
   Controller();
+
+  @override
+  construct(BuildContext context) {
+    super.construct(context);
+  }
 }