+++ LabelStoreMax/lib/config/app_payment_gateways.dart
@@ -9,7 +9,7 @@ import 'package:flutter_app/bootstrap/helpers.dart';
 | PAYMENT GATEWAYS
 |
 | Configure which payment gateways you want to use.
-| Docs here: https://woosignal.com/docs/app/ios/label-storemax
+| Docs here: https://woosignal.com/docs/app/label-storemax
 |--------------------------------------------------------------------------
 */