+++ LabelStoreMax/lib/resources/widgets/home_drawer_widget.dart
@@ -10,6 +10,7 @@
 
 
 import 'package:flutter/material.dart';
+import 'package:flutter_app/bootstrap/app_helper.dart';
 import 'package:flutter_app/bootstrap/helpers.dart';
 import 'package:flutter_app/bootstrap/shared_pref/sp_auth.dart';
 import 'package:flutter_app/resources/widgets/app_version_widget.dart';
@@ -17,6 +18,7 @@ import 'package:flutter_app/resources/widgets/woosignal_ui.dart';
 import 'package:nylo_framework/theme/helper/ny_theme.dart';
 import 'package:nylo_support/helpers/helper.dart';
 import 'package:woosignal/models/response/woosignal_app.dart';
+import 'package:url_launcher/url_launcher.dart';
 
 class HomeDrawerWidget extends StatefulWidget {
   const HomeDrawerWidget({Key key, @required this.wooSignalApp})
@@ -29,6 +31,16 @@ class HomeDrawerWidget extends StatefulWidget {
 }
 
 class _HomeDrawerWidgetState extends State<HomeDrawerWidget> {
+
+  Map<String, dynamic> _socialLinks = {};
+
+
+  @override
+  void initState() {
+    super.initState();
+    _socialLinks = AppHelper.instance.appConfig.socialLinks ?? {};
+  }
+
   @override
   Widget build(BuildContext context) {
     bool isDark = (Theme.of(context).brightness == Brightness.dark);
@@ -66,27 +78,27 @@ class _HomeDrawerWidgetState extends State<HomeDrawerWidget> {
               leading: Icon(Icons.shopping_cart),
               onTap: _actionCart,
             ),
-          if (widget.wooSignalApp.appTermslink != null && widget.wooSignalApp.appPrivacylink != null)
-            Padding(
-              child: Text(
-                trans("About Us"),
-                style: Theme.of(context).textTheme.subtitle2,
+            if (widget.wooSignalApp.appTermsLink != null && widget.wooSignalApp.appPrivacyLink != null)
+              Padding(
+                child: Text(
+                  trans("About Us"),
+                  style: Theme.of(context).textTheme.subtitle2,
+                ),
+                padding: EdgeInsets.only(left: 16, top: 8, bottom: 8),
               ),
-              padding: EdgeInsets.only(left: 16, top: 8, bottom: 8),
-            ),
-            if (widget.wooSignalApp.appTermslink != null &&
-                widget.wooSignalApp.appTermslink.isNotEmpty)
+            if (widget.wooSignalApp.appTermsLink != null &&
+                widget.wooSignalApp.appTermsLink.isNotEmpty)
               ListTile(
                 title: Text(trans("Terms and conditions"),
-                    style: Theme.of(context).textTheme.bodyText2.copyWith(
-                        fontSize: 16
-                    ),),
+                  style: Theme.of(context).textTheme.bodyText2.copyWith(
+                      fontSize: 16
+                  ),),
                 leading: Icon(Icons.menu_book_rounded),
                 trailing: Icon(Icons.keyboard_arrow_right_rounded),
                 onTap: _actionTerms,
               ),
-            if (widget.wooSignalApp.appPrivacylink != null &&
-                widget.wooSignalApp.appPrivacylink.isNotEmpty)
+            if (widget.wooSignalApp.appPrivacyLink != null &&
+                widget.wooSignalApp.appPrivacyLink.isNotEmpty)
               ListTile(
                 title: Text(trans("Privacy policy"), style: Theme.of(context).textTheme.bodyText2.copyWith(
                     fontSize: 16
@@ -99,7 +111,7 @@ class _HomeDrawerWidgetState extends State<HomeDrawerWidget> {
               title: Text(
                   trans((isDark ? "Light Mode" : "Dark Mode")),
                   style: Theme.of(context).textTheme.bodyText2.copyWith(
-                    fontSize: 16
+                      fontSize: 16
                   )
               ),
               leading: Icon(Icons.brightness_4_rounded),
@@ -109,6 +121,26 @@ class _HomeDrawerWidgetState extends State<HomeDrawerWidget> {
                 });
               },
             ),
+            if (_socialLinks.isNotEmpty)
+            Padding(
+              child: Text(
+                trans("Social"),
+                style: Theme.of(context).textTheme.subtitle2,
+              ),
+              padding: EdgeInsets.only(left: 16, top: 8, bottom: 8),
+            ),
+            ..._socialLinks.entries.where((element) => element != null && element.value != "").map((socialLink) => ListTile(
+              title: Text(
+                  capitalize(socialLink.key),
+                  style: Theme.of(context).textTheme.bodyText2.copyWith(
+                      fontSize: 16
+                  )
+              ),
+              leading: Image.asset(getImageAsset(socialLink.key) + '.png', height: 25, width: 25),
+              onTap: () async {
+                await launch(socialLink.value);
+              },
+            )).toList(),
             ListTile(
               title: AppVersionWidget(),
             ),
@@ -118,9 +150,9 @@ class _HomeDrawerWidgetState extends State<HomeDrawerWidget> {
     );
   }
 
-  _actionTerms() => openBrowserTab(url: widget.wooSignalApp.appTermslink);
+  _actionTerms() => openBrowserTab(url: widget.wooSignalApp.appTermsLink);
 
-  _actionPrivacy() => openBrowserTab(url: widget.wooSignalApp.appPrivacylink);
+  _actionPrivacy() => openBrowserTab(url: widget.wooSignalApp.appPrivacyLink);
 
   _actionProfile() async {
     Navigator.pop(context);
@@ -136,4 +168,6 @@ class _HomeDrawerWidgetState extends State<HomeDrawerWidget> {
     Navigator.pop(context);
     Navigator.pushNamed(context, "/cart");
   }
+
+  String capitalize(String s) => s[0].toUpperCase() + s.substring(1);
 }